Merge install-image_$(ARCH)_$(FEATURESET)_$(FLAVOUR)_plain_image rules
The filename of the kernel image to be installed, and the stem of the
installed name, varies between architectures, so we define several
different rules to install it for different sets of architectures.
However the basic fact that we need to install this file in /boot does
not.

We also duplicate this name information in gencontrol.py and in
debian/config/{armel,armhf,sh4}/defines (used by buildcheck.py).

To address this:

* Define [image]install-stem and [build]image-file for each architecture
* Copy these settings to make-flags in gencontrol.py
* Copy [image]install-stem to the image-stem template variable in
  gencontrol.py
* Replace the per-architecture rules with a single rule using those
  make-flags

The per-architecture rules for ARM and PowerPC also installed DTB
and DTS files, respectively.  Include those commands in the single
rule with appropriate conditions around them.



git-svn-id: svn://svn.debian.org/svn/kernel/dists/trunk/linux@21253 510b9475-24dd-0310-9b6c-e0eefe99d49f