Fix the mess around the CVE-2013-0228 fix
The fix for CVE-2013-0228 ('x86/xen: don't assume %ds is usable in
xen_iret for 32-bit PVOPS.') was previously applied only as part of
the Xen featureset.  Now we have it in all other Xen-aware kernels
via 2.6.32.61, but it is getting reverted for the xen featureset
(I think this was meant to avoid trying to apply it twice).

Amend the changelog entry for 2.6.32-46squeeze1 to note that only the
xen featureset was fixed then.

Drop the single patch and the revert line in
debian/patches/series/48squeeze9-extra


git-svn-id: svn://svn.debian.org/svn/kernel/dists/squeeze-security/linux-2.6@22109 510b9475-24dd-0310-9b6c-e0eefe99d49f